feat: resource path util
