+++ screens/controls.py
@@ -3,10 +3,11 @@ import pygame
 
 from .background import slow_bg_obj
 from models.icon_button import IconButton
+from utils.resource_path import resource_path
+
 from constants import WIDTH,\
     controlImage,\
     goBackImage, \
-    resource_path, \
     CANVAS, \
     soundList, \
     framespersec, \
@@ -109,10 +110,12 @@ def controls():
     center_x = screen_rect.centerx
     starting_x = center_x - background_width//2
 
-    control_title_font = pygame.font.Font(
-        os.path.join(FONT_PATH, 'edit_undo.ttf'), 50)
-    control_font = pygame.font.Font(os.path.join(FONT_PATH, 'neue.ttf'), 30)
-    keys_font = pygame.font.Font(os.path.join(FONT_PATH, 'neue.ttf'), 30)
+    control_title_font = pygame.font.Font(resource_path(
+        os.path.join(FONT_PATH, 'edit_undo.ttf')), 50)
+    control_font = pygame.font.Font(resource_path(
+        os.path.join(FONT_PATH, 'neue.ttf')), 30)
+    keys_font = pygame.font.Font(resource_path(
+        os.path.join(FONT_PATH, 'neue.ttf')), 30)
 
     go_back_btn = IconButton(goBackImage, (starting_x + 30, 30))
