+++ screens/score_board.py
@@ -1,14 +1,15 @@
+import os
+import pygame
+
 from models.icon_button import IconButton
 from .background import slow_bg_obj
 from .controls import audio_cfg, display_cfg
-import os
-import pygame
+from utils.resource_path import resource_path
 
 from constants import WIDTH, \
     CANVAS, \
     goBackImage, \
     center_x, \
-    resource_path, \
     score_list, \
     trophyImage, \
     framespersec, \
